Move tests to their directories
